Added updateGoogleSheet to GX_Map_utils
updateGoogleSheet is a function that takes the data that is put into the reports and writes it to a master Google Sheet containing all the subjects data.  You need the json file with the credentials and to specify the path to the json file for it to work.  updateGoogleSheet gets called by genHtmlPdf because it is part of generating reports and is reliant on the same html_parameters dictionary